genomic-service/requirements.txt
Django==5.1.0
djangorestframework==3.15.2
django-cors-headers==4.3.1
drf-spectacular==0.27.2
pymongo==4.6.0
